+++ src/components/DBSComponent/index.module.css
@@ -1,3 +1,3 @@
-.DBSUploader {
+.DBSComponent {
   padding: 15px;
 }
